Refactor handling of admins
Now we process all boundery=administrative with admin_level<9 and put it into Postgis database, this allows us check if polygon is valid/selfintersecting, but also keep track number of specific admin_level relations per country. In future we can also add checking overlapping of relations on same admin_level...

+++ PbfParser/OsmBinary/BinarySerializer.cs
@@ -551,11 +551,11 @@ namespace OsmSharp.IO.Binary
         {
             if (value == null)
             {
-                stream.WriteVarInt32((int)0);
+                stream.WriteVarInt32(0);
             }
             else if (string.IsNullOrWhiteSpace(value))
             {
-                stream.WriteVarInt32((int)1);
+                stream.WriteVarInt32(1);
             }
             else
             {
